- Using Gate::authorize() inside methods reduces unnecessary checks.
- move the middleware to routes

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -15,12 +16,12 @@ class AttendeeController extends Controller
 
     private array $relations = ['user'];
 
-    public function __construct()
-    {
-        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
-        $this->middleware('throttle:api')->only(['store', 'destroy']);
-        $this->authorizeResource(Attendee::class, 'attendee');
-    }
+    // public function __construct()
+    // {
+    //     $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    //     $this->middleware('throttle:api')->only(['store', 'destroy']);
+    //     $this->authorizeResource(Attendee::class, 'attendee');
+    // }
 
     protected function getRelations(): array
     {
@@ -31,6 +32,7 @@ class AttendeeController extends Controller
      */
     public function index(Event $event)
     {
+        Gate::authorize('viewAny', Attendee::class);
         $attendees = $this->loadRelationships(
             $event->attendees()->latest()
         );
@@ -45,6 +47,7 @@ class AttendeeController extends Controller
      */
     public function store(Request $request, Event $event)
     {
+        Gate::authorize('create', Attendee::class);
         $attendee = $this->loadRelationships(
             $event->attendees()->create([
                 'user_id' => $request->user()->id
@@ -59,6 +62,7 @@ class AttendeeController extends Controller
      */
     public function show(Event $event, Attendee $attendee)
     {
+        Gate::authorize('view', $attendee);
         return new AttendeeResource(
             $this->loadRelationships($attendee)
         );
@@ -77,6 +81,7 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+        Gate::authorize('delete', $attendee);
         $attendee->delete();
 
         return response(status: 204);
